fp-stack-docker/requirements.txt
python-neutronclient
python-novaclient
python-glanceclient
python-cinderclient
python-keystoneclient
keystoneauth1
docker
